Replace react quill and sanitize strings on previews
