+++ package.json
@@ -55,6 +55,7 @@
     "@storybook/addon-info": "^5.3.19",
     "@storybook/addon-knobs": "^5.3.19",
     "@storybook/react": "^5.3.19",
+    "@types/dompurify": "^2.0.2",
     "@types/lodash": "^4.14.155",
     "@types/react": "^16.9.36",
     "@types/react-dates": "^17.1.12",
@@ -66,6 +67,7 @@
     "autoprefixer": "^9.8.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
+    "dompurify": "^2.0.11",
     "eslint": "^7.2.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-config-prettier": "^6.11.0",
@@ -77,6 +79,7 @@
     "eslint-plugin-react": "^7.20.0",
     "eslint-plugin-react-hooks": "^4.0.4",
     "eslint-watch": "^7.0.0",
+    "htmr": "^0.8.8",
     "lodash-es": "^4.17.15",
     "moment": "^2.26.0",
     "polished": "^3.6.4",
